requirements.txt
# Core dependencies
pandas>=1.5.0,<2.0.0  # Use 1.x for compatibility
openpyxl>=3.0.9  # For Excel file reading
xlrd>=2.0.1  # For older .xls files
PyYAML>=6.0
requests>=2.26.0
beautifulsoup4>=4.10.0  # For web scraping
lxml>=4.9.0  # For BeautifulSoup parsing

# Git integration
GitPython>=3.1.24

# Data processing
numpy>=1.21.0
python-dateutil>=2.8.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.2
plotly>=5.10.0
networkx>=2.6.3

# Web interfaces
streamlit>=1.25.0  # For interactive browser
fastapi>=0.100.0  # For REST API
uvicorn>=0.23.0  # ASGI server

# Utilities
tqdm>=4.65.0  # Progress bars
click>=8.1.0  # CLI interface
python-dotenv>=1.0.0  # Environment variables
jsonschema>=4.0.0  # JSON/YAML validation
tabulate>=0.9.0  # Pretty tables